refactor: use uint64 for validator index mappings and fix forge fmt
Amp-Thread-ID: https://ampcode.com/threads/T-019c6969-b5f5-70bf-b322-0da36b36f807

+++ tips/ref-impls/src/ValidatorConfigV2.sol
@@ -30,10 +30,10 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
     Validator[] internal validatorsArray;
 
     /// @dev 1-indexed: 0 means not found. Stored value is arrayIndex + 1.
-    mapping(address => uint256) internal addressToIndex;
+    mapping(address => uint64) internal addressToIndex;
 
     /// @dev 1-indexed: 0 means not found. Stored value is arrayIndex + 1.
-    mapping(bytes32 => uint256) internal pubkeyToIndex;
+    mapping(bytes32 => uint64) internal pubkeyToIndex;
 
     uint64 internal nextDkgCeremony;
 
@@ -99,7 +99,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
             revert Unauthorized();
         }
 
-        uint256 idx = addressToIndex[validatorAddress];
+        uint64 idx = addressToIndex[validatorAddress];
         if (idx == 0) {
             revert ValidatorNotFound();
         }
@@ -141,7 +141,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
         onlyInitialized
         onlyOwnerOrValidator(validatorAddress)
     {
-        uint256 idx = addressToIndex[validatorAddress];
+        uint64 idx = addressToIndex[validatorAddress];
         if (idx == 0) {
             revert ValidatorNotFound();
         }
@@ -179,7 +179,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
             revert Unauthorized();
         }
 
-        uint256 idx = addressToIndex[validatorAddress];
+        uint64 idx = addressToIndex[validatorAddress];
         if (idx == 0) {
             revert ValidatorNotFound();
         }
@@ -210,7 +210,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
             revert InvalidValidatorAddress();
         }
 
-        uint256 idx = addressToIndex[currentAddress];
+        uint64 idx = addressToIndex[currentAddress];
         if (idx == 0) {
             revert ValidatorNotFound();
         }
@@ -239,10 +239,10 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
 
     /// @inheritdoc IValidatorConfigV2
     function getActiveValidators() external view returns (Validator[] memory validators) {
-        uint256 len = validatorsArray.length;
+        uint64 len = uint64(validatorsArray.length);
         validators = new Validator[](len);
-        uint256 idx = 0;
-        for (uint256 i = 0; i < len; i++) {
+        uint64 idx = 0;
+        for (uint64 i = 0; i < len; i++) {
             Validator storage v = validatorsArray[i];
             if (v.deactivatedAtHeight == 0) {
                 validators[idx] = v;
@@ -268,7 +268,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
     }
 
     /// @inheritdoc IValidatorConfigV2
-    function validatorByIndex(uint256 index) external view returns (Validator memory) {
+    function validatorByIndex(uint64 index) external view returns (Validator memory) {
         if (index >= validatorsArray.length) {
             revert ValidatorNotFound();
         }
@@ -277,7 +277,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
 
     /// @inheritdoc IValidatorConfigV2
     function validatorByAddress(address validatorAddress) external view returns (Validator memory) {
-        uint256 idx = addressToIndex[validatorAddress];
+        uint64 idx = addressToIndex[validatorAddress];
         if (idx == 0) {
             revert ValidatorNotFound();
         }
@@ -286,7 +286,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
 
     /// @inheritdoc IValidatorConfigV2
     function validatorByPublicKey(bytes32 publicKey) external view returns (Validator memory) {
-        uint256 idx = pubkeyToIndex[publicKey];
+        uint64 idx = pubkeyToIndex[publicKey];
         if (idx == 0) {
             revert ValidatorNotFound();
         }
@@ -383,7 +383,7 @@ contract ValidatorConfigV2 is IValidatorConfigV2 {
             revert InvalidValidatorAddress();
         }
         // Allow reusing addresses of deactivated validators
-        uint256 idx1 = addressToIndex[validatorAddress];
+        uint64 idx1 = addressToIndex[validatorAddress];
         if (idx1 != 0 && validatorsArray[idx1 - 1].deactivatedAtHeight == 0) {
             revert AddressAlreadyHasValidator();
         }
